Removed temporary patch for bug 1358192
git-svn-id: https://sdcc.svn.sourceforge.net/svnroot/sdcc/trunk/sdcc@3959 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ device/lib/printf_large.c
@@ -211,7 +211,7 @@ static int output_float (float f, unsigned char reqWidth,
 
   // split the float
   integerPart=f;
-  decimalPart=f-(float)integerPart;
+  decimalPart=f-integerPart;
 
   // fill the buffer with the integerPart (in reversed order!)
   while (integerPart) {
@@ -235,7 +235,7 @@ static int output_float (float f, unsigned char reqWidth,
       // truncate the float
       integerPart=decimalPart;
       fpBuffer[fpBD++]='0' + integerPart;
-      decimalPart-=(float)integerPart;
+      decimalPart-=integerPart;
     } while (--i);
   }
